A tank game between two players using processing's network capabilities.
This program uses networking to create a tank game between two users.
The tanks are controlled through key clicks and the shooting of the
missiles
are controlled via mouse.
The bricks are gradually created to stand in between the two players,
and
upon breaking of brick, it spawns a powerup to the player that did the
last hit.
This is the client, and requires the server to run for full
functionality.